rename file to textFormatting
