+++ src/utils/textFormatting.ts
